backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow>=10.2.0
django-filter==23.5
gunicorn==21.2.0
psycopg[binary]==3.3.3
whitenoise==6.6.0

